# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Documentation",
        "description": "Documentation for the Car Reservation API",
        "contact": {
            "email": "admin@example.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/api/user/register": {
            "post": {
                "tags": [
                    "User"
                ],
                "summary": "Registrasi pengguna baru",
                "operationId": "0234738a28358a02b0cfe4f8234a85bb",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "John Doe"
                                    },
                                    "email": {
                                        "type": "string",
                                        "example": "john@example.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "secret123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "User Berhasil Dibuat"
                    },
                    "500": {
                        "description": "Kesalahan Server"
                    }
                }
            }
        },
        "/api/user/{id}": {
            "get": {
                "tags": [
                    "User"
                ],
                "summary": "Ambil data pengguna berdasarkan ID",
                "operationId": "a826e4b095cd58b7033d53e228ae24ad",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Data pengguna ditemukan"
                    }
                }
            },
            "put": {
                "tags": [
                    "User"
                ],
                "summary": "Perbarui data pengguna",
                "operationId": "2d2cc8ddcc793bf58f51b92f251eeaea",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "email"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "John Doe Updated"
                                    },
                                    "email": {
                                        "type": "string",
                                        "example": "john_updated@example.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "newpassword",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Data pengguna berhasil diperbarui"
                    },
                    "404": {
                        "description": "Pengguna tidak ditemukan"
                    },
                    "500": {
                        "description": "Kesalahan server"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "User"
                ],
                "summary": "Hapus pengguna berdasarkan ID",
                "operationId": "08039ba8ddbb21fcffdece34b1c8955e",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Pengguna berhasil dihapus"
                    },
                    "404": {
                        "description": "Pengguna tidak ditemukan"
                    },
                    "500": {
                        "description": "Kesalahan server"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/user": {
            "get": {
                "tags": [
                    "User"
                ],
                "summary": "Ambil semua pengguna",
                "operationId": "0fa048f1cbe7e39fa4bc89a37b30f4ce",
                "responses": {
                    "200": {
                        "description": "Daftar semua pengguna berhasil diambil"
                    },
                    "500": {
                        "description": "Kesalahan server"
                    }
                }
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Bearer token authentication using Laravel Sanctum",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            },
            "security": [
                {
                    "sanctum": []
                }
            ]
        }
    },
...
}